[WIP] clustering refactor
