+++ src/dartsort/cluster/initial.py
@@ -13,20 +13,22 @@ combined with calls to `main.subtract()`, as implemented in the
 import h5py
 from dartsort.util.data_util import DARTsortSorting
 
-from . import cluster_util
+from . import cluster_util, ensemble_utils
+import numpy as np
 
 
 def cluster_chunk(
     peeling_hdf5_filename,
+    chunk_time_range_s=None,
     motion_est=None,
     strategy="closest_registered_channels",
-    chunk_size_s=300,
 ):
     """Cluster spikes from a single segment
 
     Arguments
     ---------
     peeling_hdf5_filename : str or Path
+    chunk_time_range_s : start and end time of chunk (seconds) in an iterable
     motion_est : optional dredge.motion_util.MotionEstimate
     strategy : one of "closest_registered_channels" or other choices tba
 
@@ -35,8 +37,7 @@ def cluster_chunk(
     sorting : DARTsortSorting
     """
     assert strategy in ("closest_registered_channels","hdbscan","ensembling_hdbscan",)
-    
-
+        
     if strategy == "closest_registered_channels":
         with h5py.File(peeling_hdf5_filename, "r") as h5:
             times_samples = h5["times_samples"][:]
@@ -45,9 +46,13 @@ def cluster_chunk(
             xyza = h5["point_source_localizations"][:]
             amps = h5["denoised_amplitudes"][:]
             geom = h5["geom"][:]
-        labels = cluster_util.closest_registered_channels(
-            times_s, xyza[:, 0], xyza[:, 2], geom, motion_est
+            
+        in_chunk = ensemble_utils.get_indices_in_chunk(times_s, chunk_time_range_s)
+        labels = -1 * np.ones(len(times_samples))
+        labels[in_chunk] = cluster_util.closest_registered_channels(
+            times_s[in_chunk], xyza[in_chunk, 0], xyza[in_chunk, 2], geom, motion_est
         )
+        #triaging here maybe
         sorting = DARTsortSorting(
             times_samples=times_samples,
             channels=channels,
@@ -66,29 +71,11 @@ def cluster_chunk(
             xyza = h5["point_source_localizations"][:]
             amps = h5["denoised_amplitudes"][:]
             geom = h5["geom"][:]
-        labels = cluster_util.hdbscan_clustering(
-            times_s, xyza[:, 0], xyza[:, 2], geom, amps, motion_est
-        )
-        sorting = DARTsortSorting(
-            times_samples=times_samples,
-            channels=channels,
-            labels=labels,
-            extra_features=dict(
-                point_source_localizations=xyza,
-                denoised_amplitudes=amps,
-                times_seconds=times_s,
-            ),
-        )
-    elif strategy == "ensembling_hdbscan":
-        with h5py.File(peeling_hdf5_filename, "r") as h5:
-            times_samples = h5["times_samples"][:]
-            channels = h5["channels"][:]
-            times_s = h5["times_seconds"][:]
-            xyza = h5["point_source_localizations"][:]
-            amps = h5["denoised_amplitudes"][:]
-            geom = h5["geom"][:]
-        labels = cluster_util.ensembling_hdbscan(
-            times_s, xyza[:, 0], xyza[:, 2], geom, amps, motion_est, chunk_size_s,
+            
+        in_chunk = ensemble_utils.get_indices_in_chunk(times_s, chunk_time_range_s)
+        labels = -1 * np.ones(len(times_samples))
+        labels[in_chunk]  = cluster_util.hdbscan_clustering(
+            times_s[in_chunk] , xyza[in_chunk, 0], xyza[in_chunk, 2], geom, amps[in_chunk], motion_est
         )
         sorting = DARTsortSorting(
             times_samples=times_samples,
@@ -102,5 +89,64 @@ def cluster_chunk(
         )
     else:
         raise ValueError
+    return sorting
+
+def ensemble_chunks(
+    peeling_hdf5_filename,
+    chunk_size_s=300,
+    motion_est=None,
+    ensemble_strategy="forward_backward",
+    cluster_strategy="closest_registered_channels",
+):
+    """Cluster spikes from a single segment
+
+    Arguments
+    ---------
+    peeling_hdf5_filename : str or Path
+    chunk_size_s : time in seconds for each chunk to be clustered and ensembled
+    motion_est : optional dredge.motion_util.MotionEstimate
+    ensemble_strategy : one of "forward_backward" or other choices tba
+    cluster_strategy : one of "closest_registered_channels" or other choices tba
+    
+    Returns
+    -------
+      : DARTsortSorting
+    """
+    
+    if ensemble_strategy is None:
+        #cluster full chunk with no ensembling
+        sorting = cluster_chunk(peeling_hdf5_filename, 
+                                chunk_time_range_s=None,
+                                motion_est=motion_est,
+                                strategy=cluster_strategy,
+                            )
+    else:
+        assert ensemble_strategy in ("forward_backward","meet")
+        #for loop cluster chunks
+        if ensemble_strategy == "forward_backward":
+            with h5py.File(peeling_hdf5_filename, "r") as h5:
+                times_samples = h5["times_samples"][:]
+                channels = h5["channels"][:]
+                times_s = h5["times_seconds"][:]
+                xyza = h5["point_source_localizations"][:]
+                amps = h5["denoised_amplitudes"][:]
+                geom = h5["geom"][:]
+            labels = ensemble_utils.ensembling_hdbscan(
+                times_s, xyza[:, 0], xyza[:, 2], geom, amps, motion_est, chunk_size_s,
+            )
+            sorting = DARTsortSorting(
+                times_samples=times_samples,
+                channels=channels,
+                labels=labels,
+                extra_features=dict(
+                    point_source_localizations=xyza,
+                    denoised_amplitudes=amps,
+                    times_seconds=times_s,
+                ),
+            )
+        if ensemble_strategy == "meet":
+            raise ValueError("WIP")
+        else:
+            raise ValueError
     return sorting
